requirements.txt
grpcio
grpcio-tools
joblib
pandas
scikit-learn
redis
